Add `updateLens.js` and `zipWith`
