+++ lib/updateLens/index.js
@@ -1,6 +1,4 @@
 import curry3 from '../_internal/_curry3'
+import updateLens from './updateLens'
 
-export default curry3(function updateLens(lens, cb, value) {
-  var l = lens(value)
-  return l.set(cb(l.get()))
-})
+export default curry3(updateLens)
