+++ lib/updateLens/updateLens.js
@@ -1,0 +1,4 @@
+export default function updateLens(lens, cb, value) {
+  var l = lens(value)
+  return l.set(cb(l.get()))
+}
